+++ tests/unit/plugins/modules/test_goss_validate.py
@@ -27,7 +27,7 @@ def test_goss_validate_gossfile(capfd):
 
 def test_goss_validate_format_vars(capfd):
     """test goss validate with format and vars"""
-    utils.set_module_args({'format': 'json', 'format_opts': 'pretty', 'vars': 'galaxy.yml', 'gossfile': f"{str(utils.fixtures_dir())}/goss.yaml"})
+    utils.set_module_args({'format': 'json', 'format_opts': 'pretty', 'vars': f"{str(utils.fixtures_dir())}/goss.yaml", 'gossfile': f"{str(utils.fixtures_dir())}/goss.yaml"})
     with pytest.raises(SystemExit, match='1'):
         goss_validate.main()
 
@@ -42,13 +42,12 @@ def test_goss_validate_format_vars(capfd):
     assert '-o' in info['cmd']
     assert 'pretty' in info['cmd']
     assert '--vars' in info['cmd']
-    assert 'galaxy.yml' in info['cmd']
-    assert 'Error: failed while loading vars file "galaxy.yml"' in info['stdout']
+    assert 'Error: failed while loading vars file "goss.yml"' in info['stdout']
 
 
 def test_goss_validate_retry_sleep(capfd):
     """test goss validate with retry_timeout and sleep"""
-    utils.set_module_args({'retry_timeout': '6s', 'sleep': '3s', 'gossfile': str(utils.fixtures_dir() / 'goss.yaml')})
+    utils.set_module_args({'retry_timeout': '6s', 'sleep': '3s', 'gossfile': f"{str(utils.fixtures_dir())}/goss.yaml"})
     with pytest.raises(SystemExit, match='1'):
         goss_validate.main()