+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Improve `changed` detection for `terraform_fmt` module.
 - Add `replace` parameter to `terraform_apply` module.
 - Add `diff` and `write` parameters to `packer_fmt` module.
+- Fix `vars` parameter in `goss` modules.
 
 ### 1.2.0
 - Finish adding Terraform modules.